changed the datatype of Clothing.Price from REAL/Float to Integer

+++ docs/information_retrieval/inc/table_design/table_design.sql
@@ -6,7 +6,7 @@ CREATE TABLE IF NOT EXISTS Clothing
     clothing_id     INTEGER PRIMARY KEY,
     image_name      TEXT UNIQUE NOT NULL,
     brand           TEXT,
-    price           REAL,
+    price           INTEGER,
     cloth_type      TEXT
 );
